+++ rmf_task_ros2/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog for package rmf_task_ros2
 
+2.1.0 (2022-XX-YY)
+------------------
+* ws broadcast client in dispatcher node [#212](https://github.com/open-rmf/rmf_ros2/pull/212)
+* create unique task_id with timestamp [#223](https://github.com/open-rmf/rmf_ros2/pull/223)
+
 2.0.0 (2022-03-18)
 ------------------
 No changes yet